requirements.txt
# Core
torch>=2.1.0
transformers>=4.38.0
peft>=0.9.0
datasets>=2.17.0
accelerate>=0.27.0

# Training & Evaluation
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0

# Experiment Tracking
wandb>=0.16.0

# Config
pyyaml>=6.0.1

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Utils
tqdm>=4.66.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
